满明星，初步完成登录功能

+++ src/main/java/com/cake/service/inteerfaces/IGoodService.java
@@ -61,6 +61,6 @@ public interface IGoodService {
    */
     String mincartGoodSrevice(HttpSession session);
 
-    Double getTotalPrice(HttpServletRequest request);
+    Integer getTotalPrice(HttpServletRequest request);
 
 }
